chore: make the RequestCounter class injectable and use it
