+++ .vscode/settings.json
@@ -1,1 +1,3 @@
-{}
+{
+	"cSpell.words": ["requestcounter"]
+}
