+++ src/ioc/ioc-bind.ts
@@ -5,6 +5,7 @@ import DBConnection from '../config/dbConnect';
 import CpfServices from '../services/cpfServices';
 import TYPES from './ioc-types';
 import { Configuration } from '../config/configFactory';
+import RequestCounter from '../services/requestCounter';
 
 function myContainer() {
 	return new ContainerModule((bind) => {
@@ -12,6 +13,7 @@ function myContainer() {
 		bind(Configuration).to(Configuration).inSingletonScope();
 		bind('DBConnection').to(DBConnection).inSingletonScope();
 		bind(Server).toSelf().inSingletonScope();
+		bind(RequestCounter).toSelf().inSingletonScope();
 	});
 }
