+++ src/ioc/ioc-types.ts
@@ -5,6 +5,7 @@ const TYPES = {
 	DBConnection: Symbol.for('DBConnection'),
 	Server: Symbol.for('Server'),
 	Configuration: Symbol.for('Configuration'),
+	RequestCounter: Symbol.for('RequestCounter'),
 };
 
 export default TYPES;
