+++ src/services/requestCounter.ts
@@ -1,0 +1,16 @@
+import { injectable } from 'inversify';
+
+@injectable()
+class RequestCounter {
+	private count = 0;
+
+	public increment(): void {
+		this.count++;
+	}
+
+	public getCount(): number {
+		return this.count;
+	}
+}
+
+export default RequestCounter;
